move and modify parse_boolean_arg
